+++ PostHog/SwiftUI/PostHogTagViewModifier.swift
@@ -297,15 +297,13 @@
         }
 
         func makeUIView(context: Context) -> PostHogTagUIView {
-            let view = PostHogTagUIView(id: id) { controller in
+            return PostHogTagUIView(id: id) { controller in
                 let targets = getTargetViews(from: controller)
                 if !targets.isEmpty {
                     context.coordinator.cachedTargets = targets
                     onChangeHandler(targets, [])
                 }
             }
-
-            return view
         }
 
         func updateUIView(_ uiView: PostHogTagUIView, context _: Context) {