+++ PostHogExample/ContentView.swift
@@ -435,7 +435,9 @@ struct ContentView: View {
                 }
             }
             .navigationTitle("PostHog")
-        }.onAppear {
+        }
+        .postHogDeepLinkListener()
+        .onAppear {
             api.listBeers(completion: { beers in
                 api.beers = beers
             })