+++ PostHogTests/PostHogMulticastCallbackTest.swift
@@ -2,7 +2,6 @@ import Foundation
 @testable import PostHog
 import Testing
 
-@Suite("PostHogMulticastCallback Tests")
 class PostHogMulticastCallbackTests {
     @Test("Single subscriber receives value")
     func singleSubscriber() {
@@ -97,7 +96,6 @@ class PostHogMulticastCallbackTests {
     }
 }
 
-@Suite("PostHogThrottledMulticastCallback Tests")
 class PostHogThrottledMulticastCallbackTests {
     @Test("Single subscriber receives value with throttle")
     func singleSubscriber() async {